refactor OpenID4VCIContext
